Back-End Projects/PokemonProject_DockerDjango/PokemonApp/requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
django-filter==3.5.2
psycopg2==2.9.5
sqlparse==0.4.3



